+++ crates/libwebrtc-sys/src/lib.rs
@@ -1938,13 +1938,28 @@ impl PeerConnectionFactoryInterface {
         Ok(MediaStreamInterface(ptr))
     }
 
-    /// Returns the capabilities of an RTP sender of type `kind`.
+    /// Returns the capabilities of an [RTP] sender of the provided
+    /// [`MediaType`].
+    ///
+    /// [RTP]: https://en.wikipedia.org/wiki/Real-time_Transport_Protocol
     #[must_use]
     pub fn get_rtp_sender_capabilities(
         &self,
         kind: MediaType,
     ) -> RtpCapabilities {
         RtpCapabilities(webrtc::get_rtp_sender_capabilities(&self.0, kind))
+    }
+
+    /// Returns the capabilities of an [RTP] receiver of the provided
+    /// [`MediaType`].
+    ///
+    /// [RTP]: https://en.wikipedia.org/wiki/Real-time_Transport_Protocol
+    #[must_use]
+    pub fn get_rtp_receiver_capabilities(
+        &self,
+        kind: MediaType,
+    ) -> RtpCapabilities {
+        RtpCapabilities(webrtc::get_rtp_receiver_capabilities(&self.0, kind))
     }
 }