+++ crates/native/src/frb_generated.rs
@@ -38,7 +38,7 @@ flutter_rust_bridge::frb_generated_boilerplate!(
     default_rust_auto_opaque = RustAutoOpaqueMoi,
 );
 pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_VERSION: &str = "2.4.0";
-pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = -602564949;
+pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = -1353945569;
 
 // Section: executor
 
@@ -613,6 +613,40 @@ fn wire__crate__api__get_peer_stats_impl(
                         Ok(output_ok)
                     })(),
                 )
+            }
+        },
+    )
+}
+fn wire__crate__api__get_rtp_receiver_capabilities_impl(
+    port_: flutter_rust_bridge::for_generated::MessagePort,
+    ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
+    rust_vec_len_: i32,
+    data_len_: i32,
+) {
+    FLUTTER_RUST_BRIDGE_HANDLER.wrap_normal::<flutter_rust_bridge::for_generated::SseCodec, _, _>(
+        flutter_rust_bridge::for_generated::TaskInfo {
+            debug_name: "get_rtp_receiver_capabilities",
+            port: Some(port_),
+            mode: flutter_rust_bridge::for_generated::FfiCallMode::Normal,
+        },
+        move || {
+            let message = unsafe {
+                flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(
+                    ptr_,
+                    rust_vec_len_,
+                    data_len_,
+                )
+            };
+            let mut deserializer =
+                flutter_rust_bridge::for_generated::SseDeserializer::new(message);
+            let api_kind = <crate::api::MediaType>::sse_decode(&mut deserializer);
+            deserializer.end();
+            move |context| {
+                transform_result_sse::<_, ()>((move || {
+                    let output_ok =
+                        Result::<_, ()>::Ok(crate::api::get_rtp_receiver_capabilities(api_kind))?;
+                    Ok(output_ok)
+                })())
             }
         },
     )
@@ -3269,42 +3303,45 @@ fn pde_ffi_dispatcher_primary_impl(
         13 => wire__crate__api__enumerate_displays_impl(port, ptr, rust_vec_len, data_len),
         14 => wire__crate__api__get_media_impl(port, ptr, rust_vec_len, data_len),
         15 => wire__crate__api__get_peer_stats_impl(port, ptr, rust_vec_len, data_len),
-        16 => wire__crate__api__get_rtp_sender_capabilities_impl(port, ptr, rust_vec_len, data_len),
-        17 => wire__crate__api__get_transceiver_direction_impl(port, ptr, rust_vec_len, data_len),
-        18 => wire__crate__api__get_transceiver_mid_impl(port, ptr, rust_vec_len, data_len),
-        19 => wire__crate__api__get_transceivers_impl(port, ptr, rust_vec_len, data_len),
-        20 => wire__crate__api__is_fake_media_impl(port, ptr, rust_vec_len, data_len),
-        21 => wire__crate__api__microphone_volume_impl(port, ptr, rust_vec_len, data_len),
-        22 => {
+        16 => {
+            wire__crate__api__get_rtp_receiver_capabilities_impl(port, ptr, rust_vec_len, data_len)
+        }
+        17 => wire__crate__api__get_rtp_sender_capabilities_impl(port, ptr, rust_vec_len, data_len),
+        18 => wire__crate__api__get_transceiver_direction_impl(port, ptr, rust_vec_len, data_len),
+        19 => wire__crate__api__get_transceiver_mid_impl(port, ptr, rust_vec_len, data_len),
+        20 => wire__crate__api__get_transceivers_impl(port, ptr, rust_vec_len, data_len),
+        21 => wire__crate__api__is_fake_media_impl(port, ptr, rust_vec_len, data_len),
+        22 => wire__crate__api__microphone_volume_impl(port, ptr, rust_vec_len, data_len),
+        23 => {
             wire__crate__api__microphone_volume_is_available_impl(port, ptr, rust_vec_len, data_len)
         }
-        23 => wire__crate__api__register_track_observer_impl(port, ptr, rust_vec_len, data_len),
-        24 => wire__crate__api__restart_ice_impl(port, ptr, rust_vec_len, data_len),
-        25 => wire__crate__api__sender_get_parameters_impl(port, ptr, rust_vec_len, data_len),
-        26 => wire__crate__api__sender_replace_track_impl(port, ptr, rust_vec_len, data_len),
-        27 => wire__crate__api__sender_set_parameters_impl(port, ptr, rust_vec_len, data_len),
-        28 => wire__crate__api__set_audio_level_observer_enabled_impl(
+        24 => wire__crate__api__register_track_observer_impl(port, ptr, rust_vec_len, data_len),
+        25 => wire__crate__api__restart_ice_impl(port, ptr, rust_vec_len, data_len),
+        26 => wire__crate__api__sender_get_parameters_impl(port, ptr, rust_vec_len, data_len),
+        27 => wire__crate__api__sender_replace_track_impl(port, ptr, rust_vec_len, data_len),
+        28 => wire__crate__api__sender_set_parameters_impl(port, ptr, rust_vec_len, data_len),
+        29 => wire__crate__api__set_audio_level_observer_enabled_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        29 => wire__crate__api__set_audio_playout_device_impl(port, ptr, rust_vec_len, data_len),
-        30 => wire__crate__api__set_codec_preferences_impl(port, ptr, rust_vec_len, data_len),
-        31 => wire__crate__api__set_local_description_impl(port, ptr, rust_vec_len, data_len),
-        32 => wire__crate__api__set_microphone_volume_impl(port, ptr, rust_vec_len, data_len),
-        33 => wire__crate__api__set_on_device_changed_impl(port, ptr, rust_vec_len, data_len),
-        34 => wire__crate__api__set_remote_description_impl(port, ptr, rust_vec_len, data_len),
-        35 => wire__crate__api__set_track_enabled_impl(port, ptr, rust_vec_len, data_len),
-        36 => wire__crate__api__set_transceiver_direction_impl(port, ptr, rust_vec_len, data_len),
-        37 => wire__crate__api__set_transceiver_recv_impl(port, ptr, rust_vec_len, data_len),
-        38 => wire__crate__api__set_transceiver_send_impl(port, ptr, rust_vec_len, data_len),
-        39 => wire__crate__api__stop_transceiver_impl(port, ptr, rust_vec_len, data_len),
-        40 => wire__crate__api__track_height_impl(port, ptr, rust_vec_len, data_len),
-        41 => wire__crate__api__track_state_impl(port, ptr, rust_vec_len, data_len),
-        42 => wire__crate__api__track_width_impl(port, ptr, rust_vec_len, data_len),
-        43 => wire__crate__api__video_decoders_impl(port, ptr, rust_vec_len, data_len),
-        44 => wire__crate__api__video_encoders_impl(port, ptr, rust_vec_len, data_len),
+        30 => wire__crate__api__set_audio_playout_device_impl(port, ptr, rust_vec_len, data_len),
+        31 => wire__crate__api__set_codec_preferences_impl(port, ptr, rust_vec_len, data_len),
+        32 => wire__crate__api__set_local_description_impl(port, ptr, rust_vec_len, data_len),
+        33 => wire__crate__api__set_microphone_volume_impl(port, ptr, rust_vec_len, data_len),
+        34 => wire__crate__api__set_on_device_changed_impl(port, ptr, rust_vec_len, data_len),
+        35 => wire__crate__api__set_remote_description_impl(port, ptr, rust_vec_len, data_len),
+        36 => wire__crate__api__set_track_enabled_impl(port, ptr, rust_vec_len, data_len),
+        37 => wire__crate__api__set_transceiver_direction_impl(port, ptr, rust_vec_len, data_len),
+        38 => wire__crate__api__set_transceiver_recv_impl(port, ptr, rust_vec_len, data_len),
+        39 => wire__crate__api__set_transceiver_send_impl(port, ptr, rust_vec_len, data_len),
+        40 => wire__crate__api__stop_transceiver_impl(port, ptr, rust_vec_len, data_len),
+        41 => wire__crate__api__track_height_impl(port, ptr, rust_vec_len, data_len),
+        42 => wire__crate__api__track_state_impl(port, ptr, rust_vec_len, data_len),
+        43 => wire__crate__api__track_width_impl(port, ptr, rust_vec_len, data_len),
+        44 => wire__crate__api__video_decoders_impl(port, ptr, rust_vec_len, data_len),
+        45 => wire__crate__api__video_encoders_impl(port, ptr, rust_vec_len, data_len),
         _ => unreachable!(),
     }
 }