+++ ios/Classes/controller/PeerConnectionFactoryController.swift
@@ -63,6 +63,15 @@ class PeerConnectionFactoryController {
         .rtpSenderCapabilities(kind: mediaKind.intoWebRtc())
 
       result(capabilities.asFlutterResult())
+    case "getRtpReceiverCapabilities":
+      let kind = argsMap!["kind"] as? Int
+      let mediaKind = MediaType(rawValue: kind!)!
+
+      let capabilities = self
+        .peerFactory
+        .rtpReceiverCapabilities(kind: mediaKind.intoWebRtc())
+
+      result(capabilities.asFlutterResult())
     case "videoEncoders":
       let res = [
         VideoCodecInfo(